project finished

+++ pages/api/products/[id].js
@@ -7,7 +7,9 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies
   } = req;
+  const token = cookies.token
 
   dbConnect();
   // methode get sur mongose pour recuperer les desc,title,price des produits
@@ -22,6 +24,9 @@ export default async function handler(req, res) {
   }
   // methode PUT pour la mise a jour des produits et des tarifs
   if (method === "PUT") {
+     if(!token || token !== process.env.token) {
+      return res.status(401).json("Not authentification !");
+    }
 
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
@@ -34,6 +39,9 @@ export default async function handler(req, res) {
   }
   //methode delete pour la suppression  des produits et des tarifs ect
   if (method === "DELETE") {
+     if(!token || token !== process.env.token) {
+      return res.status(401).json("Not authentification !");
+    }
 
     try {
       await Product.findByIdAndDelete(id);
